move: cagetory를 post 하위로 이동
