+++ lib/features/post/data/data_source/create_post_data_source.dart
@@ -1,5 +1,5 @@
 import 'package:sodong_app/features/post/data/dto/post_dto.dart';
-import 'package:sodong_app/features/post_list/domain/models/category.dart';
+import 'package:sodong_app/features/post/domain/entities/category.dart';
 
 abstract interface class CreatePostDataSource {
   /// Post 저장
