+++ lib/features/post/data/data_source/remote_create_post_data_source.dart
@@ -3,7 +3,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:sodong_app/features/post/data/data_source/create_post_data_source.dart';
 import 'package:sodong_app/features/post/data/dto/post_dto.dart';
-import 'package:sodong_app/features/post_list/domain/models/category.dart';
+import 'package:sodong_app/features/post/domain/entities/category.dart';
 
 class RemoteCreatePostDataSource implements CreatePostDataSource {
   RemoteCreatePostDataSource(this._firestore, this._storage);
