+++ lib/features/post/data/repositories/post_detail_repository_impl.dart
@@ -2,7 +2,7 @@ import 'package:sodong_app/features/post/data/data_source/post_detail_data_sourc
 import 'package:sodong_app/features/post/data/mapper/post_mapper.dart';
 import 'package:sodong_app/features/post/domain/entities/post.dart';
 import 'package:sodong_app/features/post/domain/repository/post_detail_repository.dart';
-import 'package:sodong_app/features/post_list/domain/models/category.dart';
+import 'package:sodong_app/features/post/domain/entities/category.dart';
 
 class PostDetailRepositoryImpl implements PostDetailRepository {
   PostDetailRepositoryImpl(this.remoteDataSource, this._postMapper);
