+++ lib/features/post/domain/repository/post_detail_repository.dart
@@ -1,5 +1,5 @@
 import 'package:sodong_app/features/post/domain/entities/post.dart';
-import 'package:sodong_app/features/post_list/domain/models/category.dart';
+import 'package:sodong_app/features/post/domain/entities/category.dart';
 
 abstract class PostDetailRepository {
   Stream<Post> getPostDetail(
