+++ lib/features/post/domain/repository/post_repository.dart
@@ -1,6 +1,6 @@
 import 'package:sodong_app/core/result/result.dart';
 import 'package:sodong_app/features/post/domain/entities/post.dart';
-import 'package:sodong_app/features/post_list/domain/models/category.dart';
+import 'package:sodong_app/features/post/domain/entities/category.dart';
 
 abstract interface class PostRepository {
   /// Post 저장
