+++ lib/features/post/domain/use_case/create_post_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:sodong_app/core/result/result.dart';
 import 'package:sodong_app/features/post/domain/entities/post.dart';
 import 'package:sodong_app/features/post/domain/repository/post_repository.dart';
-import 'package:sodong_app/features/post_list/domain/models/category.dart';
+import 'package:sodong_app/features/post/domain/entities/category.dart';
 
 class CreatePostUseCase {
   CreatePostUseCase(this._repository);
