+++ lib/features/post/domain/use_case/fetch_post_detail_usecase.dart
@@ -1,6 +1,6 @@
 import 'package:sodong_app/features/post/domain/entities/post.dart';
 import 'package:sodong_app/features/post/domain/repository/post_detail_repository.dart';
-import 'package:sodong_app/features/post_list/domain/models/category.dart';
+import 'package:sodong_app/features/post/domain/entities/category.dart';
 
 class FetchPostDetailUseCase {
   FetchPostDetailUseCase(this.repository);
